conversation_to_md/requirements.txt
streamlit>=1.28.0
ijson>=3.2.0
pandas>=2.0.0
openpyxl>=3.1.0
python-slugify>=8.0.0
